requirements.txt
psycopg==3.2.3
python-dotenv==1.0.1